+++ src/app/(dashboard)/(blog)/posts/(routes)/categories/components/page/heading.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import { PageHeadingWithButton } from '@/components/page-headings';
+import { CreateButton } from '@/components/buttons';
+
+import { usePostsCategoriesStore } from '../../store/posts-categories-store';
+
+export default function PostsCategoriesPageHeading() {
+  const { setCreateSheetOpen } = usePostsCategoriesStore();
+
+  return (
+    <PageHeadingWithButton heading="Categories">
+      <CreateButton onClick={() => setCreateSheetOpen(true)} />
+    </PageHeadingWithButton>
+  );
+}